Consolidate the 3 gulp build files into 1. There are 3 very small gulp build files that makes sense to consolidate into a single gulp file since it is not very large (and eleminates the duplication amongst the files).
